Init
This program allows for the testing of up to 5 values for a particular hyper-parameter. Currently certain parts of this program are hard-coded to use 5 parameter variations, and this won't be fixed until after the COVID-19 related shutdowns of the research lab and subsequent retrieval of updated versions of this program.